third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/history-traversal/same-url-expected.txt

This is a testharness.js-based test.
FAIL Test same-URL navigation and its effects on history assert_equals: expected 2 but got 1
Harness: the test ran to completion.
